More game mechanics
Made checkpoint objects actually do something, added a collidable bool to game object so we can set them to not be collidable, and added a fourth vein to infect to trigger the bullet hell, after which the checkpoint to the bone marrow area. the bullet hell area now unlocks after the 3 veins in the main area have been infected. also added a game over screen when you lose all your health